Split test methods out of the Search class.
